♻️ Use a separate component for the Select
The idea was to align it with the Text field component to make it easier to add the validation stuffs when we have landed that in Figma
