+++ libraries/core-react/src/Select/NativeSelect/Select/index.js
@@ -1,0 +1,1 @@
+export { Select } from './Select'
